Latex-suite -> vimtex

+++ .bashrc
@@ -112,7 +112,7 @@ alias l='ls -CF'
 
 alias gtr='git update && git tree'
 git-push-FORCE-REMOTE() {
-    git push --delete origin "$1" && git push --set-upstream origin "$1"
+    [ "$1" != "" ] && git push --delete origin "$1" && git push --set-upstream origin "$1"
 }
 alias git-current-branch='git rev-parse --abbrev-ref HEAD'
